Extract Create new room, use useContext to provide uuidSet
